Issues refeitas e criação de componentes para histogramas e piechart
